backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
pydantic==2.5.0
langchain
langchain-groq==0.1.0
langgraph==0.0.32
python-dotenv==1.0.0
python-multipart==0.0.6
